move rules tests under bank_rule directory
